+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/json/request/ChannelFollowRequest.kt
@@ -1,0 +1,10 @@
+package com.gitlab.kordlib.rest.json.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+class ChannelFollowRequest(
+        @SerialName("webhook_channel_id")
+        val webhookChannelId: String
+)